FE: graphical fix company registration form, fix calculate class

+++ service-logic/src/main/java/com/itechart/service/dto/CleaningTimeDto.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class CleaningTimeDto {
-    private int standardRoomCleaningTime;
-    private int springCleaningTime;
-    private int repairAndConstructionCleaningTime;
-    private int dryCarpetCleaningTime;
-    private int officeCleaningTime;
-    private int furnitureAndCoatingsCleaningTime;
-    private int industrialCleaningTime;
-    private int poolCleaningTime;
+    private Double standardRoomCleaningTime;
+    private Double springCleaningTime;
+    private Double repairAndConstructionCleaningTime;
+    private Double dryCarpetCleaningTime;
+    private Double officeCleaningTime;
+    private Double furnitureAndCoatingsCleaningTime;
+    private Double industrialCleaningTime;
+    private Double poolCleaningTime;
     private int smallRoomCleaningTime;
     private int bigRoomCleaningTime;
     private int bathroomCleaningTime;
